pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aeko-mcp"
version = "0.4.0"
description = "AEKO MCP — AI Engine Optimization tools for Claude and Codex"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "AEKO Intelligence" },
]
keywords = ["aeo", "ai-visibility", "ecommerce", "mcp", "claude", "codex"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    # mcp and sse-starlette are capped below their starlette-1.0 bump so the
    # AEKO backend (currently on fastapi 0.115.x → starlette <0.42) can embed
    # aeko-mcp without an irreconcilable starlette upgrade.
    "mcp>=1.11.0,<1.16.0",
    "sse-starlette>=1.6.1,<3.0.0",
    "httpx>=0.27.0,<1.0.0",
    "pydantic>=2.11.0,<3.0.0",
    "Pillow>=10.3.0",
]

[project.optional-dependencies]
pdf = ["PyPDF2>=3.0.0"]
docx = ["python-docx>=1.0.0"]
all = ["PyPDF2>=3.0.0", "python-docx>=1.0.0"]

[project.urls]
Homepage = "https://aeko-intelligence.com"
Repository = "https://github.com/AEKO-Intelligence/aeko-mcp"
Issues = "https://github.com/AEKO-Intelligence/aeko-mcp/issues"

[project.scripts]
aeko-mcp = "aeko_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["aeko_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/aeko_mcp",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]
